feat(migrations): update expiration columns to use bigInteger

+++ database/migrations/2025_03_08_001449_create_game_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('game', function (Blueprint $table) {
             $table->string('key')->primary();
             $table->json('data');
-            $table->integer('expiration')->nullable();
+            $table->bigInteger('expiration')->nullable();
         });
     }
